Split virtual device and basic sharding tests
