+++ test/run_tests.sh
@@ -158,7 +158,8 @@ function run_op_tests {
   run_pjrt python3 "$CDIR/pjrt/test_experimental_tpu.py"
   run_pjrt python3 "$CDIR/pjrt/test_ddp.py"
   run_pjrt python3 "$CDIR/pjrt/test_mesh_service.py"
-  run_pjrt python3 "$CDIR/test_xla_sharding.py"
+  run_pjrt python3 "$CDIR/spmd/test_xla_sharding.py"
+  run_pjrt python3 "$CDIR/spmd/test_xla_virtual_device.py"
   run_test python3 "$CDIR/test_operations_hlo.py" "$@" --verbosity=$VERBOSITY
   run_test python3 "$CDIR/test_input_output_aliases.py"
 }
